requirements.txt
Django
djangorestframework
celery
django-celery-beat
django-celery-results
django-cors-headers
redis
requests
beautifulsoup4
python-environ
image
psycopg2-binary